Querying start at a specific second. passing double as a parameter instead of int (which specifiec the bit position)

+++ src/SoundFingerprinting/Audio/Bass/BassAudioService.cs
@@ -148,7 +148,7 @@
 
             if (startAtSecond > 0)
             {
-                if (!Bass.BASS_ChannelSetPosition(stream, startAtSecond))
+                if (!Bass.BASS_ChannelSetPosition(stream, (double) startAtSecond))
                 {
                     throw new Exception(Bass.BASS_ErrorGetCode().ToString());
                 }
